backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.11.3
python-dotenv==1.0.1
validators==0.34.0
Werkzeug==3.1.3
